+++ test/core/src/entidades/Enemigo2.java
@@ -15,7 +15,7 @@ public class Enemigo2 extends Enemigo{
 		disparot = 0.0f;
 		
 	}
-	int vel  = getVelocidadX();
+	float vel  = getVelocidadX();
 	@Override
 	public void mover() {
